+++ CHANGELOG.md
@@ -43,7 +43,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for PHP 7.0 and 7.1 (#690)
 
 ### Fixed
-- Adapt the tests and testing framework to TYPO3 10LTS (#767, #777, #778, #779)
+- Adapt the tests and testing framework to TYPO3 10LTS (#767, #777, #778, #779, #780)
 - Recognize flexforms data converted to an array (#753)
 - Always display the incorrect value in the configuration check (#751)
 - Fix the `TemplateHelper` initialization (#740)